Refactor tests to integration tests to improve coverage reporting
- Move tests from `src/semantic/conversion.rs` to `tests/conversion_coverage.rs`
- Add new tests in `tests/report_coverage.rs` for `src/report.rs`
- Expose `glossa::semantic::conversion` as public to allow external testing
- Fix various coverage gaps and ensure standard iterator methods are used in tests
- Ensure compliance with `clippy::field_reassign_with_default` and `clippy::vec_init_then_push`

This change aims to clarify the "diff hit" metric in Codecov by separating test code from production code and ensuring all complex branches in the "interpreter" logic are exercised.
